3 Floats find the maximum

+++ Generics/src/com/bridgelabz/MaximumNumber.java
@@ -1,8 +1,8 @@
 package com.bridgelabz;
 
 public class MaximumNumber {
-	public static Integer findMaximumNumber(Integer x, Integer y, Integer z) {
-		Integer max = x;
+	public static Double findMaximumNumber(Double x, Double y, Double z) {
+		Double max = x;
 		if (y.compareTo(max) > 0) {
 			max = y;
 		}
